moved logger to class, fixed encoding
