+++ Library.Api/appsettings.json
@@ -1,4 +1,6 @@
 {
+  "ConnectionStrings": {
+    "library": "mongodb://localhost:27017/library" },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
